[will-squash] moves decisionlib to own library
