--- automation/taskcluster/decisionlib/common.py
@@ -1,1 +1,0 @@
-SlugId = str
